go_tutor_agent/pyproject.toml
[project]
name = "go-tutor-agent"
version = "0.1.0"
description = "Un agente de IA para aprender Go de forma personalizada."
authors = [
    {name = "JosCarRub", email = "josecarcatarubio6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi == 0.115.12",
    "uvicorn == 0.34.3",
    "smolagents == 1.17.0",
    "litellm == 1.72.1",
    "chromadb >= 1.0.13, <2.0.0", 
    "docker >= 7.1.0, <8.0.0",     
    "beautifulsoup4 >= 4.13.4, <5.0.0", 
    "python-dotenv >= 1.1.0, <2.0.0", 
    "requests (>=2.32.4,<3.0.0)", 
    "sentence-transformers (>=4.1.0,<5.0.0)" 
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"